fix: resolve CI failures and dragging issues in timeline keyframe editor
- Fixed `UIAction::TimelineAction` import error in `crates/mapmap/src/ui/editors/timeline.rs` causing the build pipeline to fail.
- Fixed volatile egui widget ID issue during drag operations by keeping a stable identifier based on `original_time` within `dragged_keyframe` state. This prevents drag-locks from failing mid-drag when backend state updates.
- Constrained keyboard shortcut (`Delete`/`Backspace`) tracking to prevent accidental keyframe deletions when users type into a general application text input.
- Ran `cargo fmt --all` to resolve Quality Gate pipeline CI formatting check failures.

+++ crates/mapmap/src/app/actions.rs
@@ -642,36 +642,64 @@ pub fn handle_ui_actions(app: &mut App) -> Result<bool> {
                         .set_active_module(Some(module_id)),
                     TimelineAction::AddKeyframe(track_name, time) => {
                         let bindings = app.state.effect_animator_mut().bindings().to_vec();
-                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name) {
-                            if let Some(val) = app.state.effect_animator_mut().get_value(binding.id).cloned() {
-                                app.state.effect_animator_mut().add_keyframe(binding.id, time, val);
+                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name)
+                        {
+                            if let Some(val) = app
+                                .state
+                                .effect_animator_mut()
+                                .get_value(binding.id)
+                                .cloned()
+                            {
+                                app.state.effect_animator_mut().add_keyframe(
+                                    binding.id,
+                                    time as f64,
+                                    val,
+                                );
                             } else {
-                                app.state.effect_animator_mut().add_keyframe(binding.id, time, mapmap_core::animation::AnimValue::Float(0.0));
+                                app.state.effect_animator_mut().add_keyframe(
+                                    binding.id,
+                                    time as f64,
+                                    mapmap_core::animation::AnimValue::Float(0.0),
+                                );
                             }
                         }
                     }
                     TimelineAction::MoveKeyframe(track_name, old_time, new_time) => {
                         let bindings = app.state.effect_animator_mut().bindings().to_vec();
-                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name) {
+                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name)
+                        {
                             // Find the old keyframe value
                             let value = {
                                 let clip = app.state.effect_animator_mut().clip();
                                 if let Some(track) = clip.get_track(&track_name) {
-                                    track.keyframes.iter().find(|(_, k)| (k.time - old_time).abs() < 0.001).map(|(_, k)| k.value.clone())
+                                    track
+                                        .keyframes
+                                        .iter()
+                                        .find(|(_, k)| (k.time - old_time).abs() < 0.001)
+                                        .map(|(_, k)| k.value.clone())
                                 } else {
                                     None
                                 }
                             };
                             if let Some(val) = value {
-                                app.state.effect_animator_mut().remove_keyframe(binding.id, old_time);
-                                app.state.effect_animator_mut().add_keyframe(binding.id, new_time, val);
+                                app.state
+                                    .effect_animator_mut()
+                                    .remove_keyframe(binding.id, old_time as f64);
+                                app.state.effect_animator_mut().add_keyframe(
+                                    binding.id,
+                                    new_time as f64,
+                                    val,
+                                );
                             }
                         }
                     }
                     TimelineAction::DeleteKeyframe(track_name, time) => {
                         let bindings = app.state.effect_animator_mut().bindings().to_vec();
-                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name) {
-                            app.state.effect_animator_mut().remove_keyframe(binding.id, time);
+                        if let Some(binding) = bindings.iter().find(|b| b.track_name == track_name)
+                        {
+                            app.state
+                                .effect_animator_mut()
+                                .remove_keyframe(binding.id, time as f64);
                         }
                     }
                 }
